extract SSE hook and components
