+++ components/ChatArea.js
@@ -1,14 +1,10 @@
 "use client";
 
 import { useState, useRef, useEffect } from "react";
-import { Button } from "@/components/ui/button";
-import { Textarea } from "@/components/ui/textarea";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Avatar, AvatarImage } from "@/components/ui/avatar";
-import { CheckCircle2, Circle, HelpCircle, Loader2, ExternalLink, Download, FileText, ArrowUp } from 'lucide-react'; // Icons for status and Loader2
-import LoadingMessage from "@/components/LoadingMessage"; // Import the LoadingMessage component
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
+import MessageList, { isDocumentMessage } from "./chat/MessageList";
+import ChatInputForm from "./chat/ChatInputForm";
+import { useChatSSE } from "@/hooks/use-chat-sse";
+import { Circle, HelpCircle } from 'lucide-react';
 import { TOOLS } from '@/lib/config/tools'; // Import TOOLS
 import { useAuth } from "./AuthProvider";
 import { initializeThread, saveMessage, subscribeToThread } from '@/lib/utils/supabase';
@@ -37,396 +33,6 @@ const workshopQuestions = [
 ];
 
 // Add a component for rendering markdown messages
-function MarkdownMessage({ content }) {
-  return (
-    <ReactMarkdown
-      //className="prose prose-sm dark:prose-invert prose-p:my-1 prose-headings:mb-2 prose-headings:mt-4 prose-pre:my-1 max-w-none" 
-      remarkPlugins={[remarkGfm]}
-      components={{
-        // Allow <a> tags to be rendered properly
-        a: ({ node, ...props }) => (
-          <a 
-            {...props} 
-            target="_blank" 
-            rel="noopener noreferrer" 
-            className="text-blue-600 hover:underline"
-          />
-        )
-      }}
-    >
-      {content}
-    </ReactMarkdown>
-  );
-}
-
-// Add a utility to extract links from message content
-function extractN8nLinks(content) {
-  if (typeof content !== 'string') return {};
-  const googleDocMatch = content.match(/View Google Doc: (https?:\/\/[^\s]+)/);
-  const pdfWebViewMatch = content.match(/View PDF: (https?:\/\/[^\s]+)/);
-  const pdfDownloadMatch = content.match(/Download PDF: (https?:\/\/[^\s]+)/);
-  return {
-    googleDocLink: googleDocMatch ? googleDocMatch[1] : null,
-    pdfWebViewLink: pdfWebViewMatch ? pdfWebViewMatch[1] : null,
-    pdfDownloadLink: pdfDownloadMatch ? pdfDownloadMatch[1] : null,
-  };
-}
-
-// Add a function to check if a message is a document message
-function isDocumentMessage(message) {
-  const hasDocumentContent = typeof message.content === 'string' && 
-    (message.content.includes('Document generated successfully') || 
-     message.content.includes('generating your document') ||
-     message.content.includes('View Google Doc') ||
-     message.content.includes('document-generation-status'));
-  
-  const hasDocumentMetadata = message.metadata?.documentLinks && 
-    Object.values(message.metadata.documentLinks).some(link => link);
-  
-  const result = hasDocumentContent || hasDocumentMetadata;
-  
-  console.log('[isDocumentMessage] Checking message:', {
-    messageId: message.id,
-    hasDocumentContent,
-    hasDocumentMetadata,
-    contentPreview: typeof message.content === 'string' ? message.content.substring(0, 100) + '...' : 'non-string',
-    metadata: message.metadata,
-    result
-  });
-  
-  return result;
-}
-
-// Function to render HTML content directly
-function HTMLContent({ content }) {
-  // Clean up content by removing redundant "Link:" text that follows HTML links
-  const cleanContent = (rawContent) => {
-    if (!rawContent) return '';
-    
-    // Remove "Link: [URL]" patterns that appear after HTML links
-    let cleaned = rawContent.replace(/\n\nLink:\s*https?:\/\/[^\s]+/g, '');
-    
-    // Also remove just "Link:" followed by URL on same line
-    cleaned = cleaned.replace(/Link:\s*https?:\/\/[^\s]+/g, '');
-    
-    return cleaned.trim();
-  };
-  
-  // Parse the content to extract links and render them as proper React components
-  const processContent = () => {
-    const cleanedContent = cleanContent(content);
-    if (!cleanedContent) return '';
-    
-    try {
-      // Create a temporary div to parse the HTML
-      const tempDiv = document.createElement('div');
-      tempDiv.innerHTML = cleanedContent;
-      
-      // Extract all links
-      const links = [];
-      const linkElements = tempDiv.querySelectorAll('a');
-      linkElements.forEach((link, index) => {
-        const href = link.getAttribute('href');
-        const text = link.textContent;
-        const isDownload = link.hasAttribute('download');
-        
-        links.push({
-          href,
-          text,
-          isDownload,
-          index,
-          outerHTML: link.outerHTML // Store the original HTML for comparison
-        });
-      });
-      
-      // If no links found, just return the content as is
-      if (links.length === 0) {
-        return <span>{cleanedContent.replace(/<[^>]*>/g, '')}</span>;
-      }
-      
-      // Split content at link positions
-      let remainingContent = cleanedContent;
-      const fragments = [];
-      
-      links.forEach((link, i) => {
-        const linkIndex = remainingContent.indexOf(link.outerHTML);
-        if (linkIndex >= 0) {
-          // Add text before the link
-          if (linkIndex > 0) {
-            fragments.push({
-              type: 'text',
-              content: remainingContent.substring(0, linkIndex).replace(/<[^>]*>/g, ''),
-              key: `text-${i}`
-            });
-          }
-          
-          // Add the link
-          fragments.push({
-            type: 'link',
-            href: link.href,
-            text: link.text,
-            isDownload: link.isDownload,
-            key: `link-${i}`
-          });
-          
-          // Update remaining content
-          remainingContent = remainingContent.substring(linkIndex + link.outerHTML.length);
-        }
-      });
-      
-      // Add any remaining text
-      if (remainingContent) {
-        fragments.push({
-          type: 'text',
-          content: remainingContent.replace(/<[^>]*>/g, ''),
-          key: `text-final`
-        });
-      }
-      
-      // Render the fragments
-      return (
-        <div className="space-y-2">
-          {fragments.map(fragment => {
-            if (fragment.type === 'text') {
-              return <span key={fragment.key}>{fragment.content}</span>;
-            } else {
-              return (
-                <a 
-                  key={fragment.key}
-                  href={fragment.href}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  download={fragment.isDownload}
-                  className="text-blue-600 hover:underline font-medium"
-                >
-                  {fragment.text}
-                </a>
-              );
-            }
-          })}
-        </div>
-      );
-    } catch (error) {
-      console.error('Error processing HTML content:', error);
-      // Fallback to simply removing HTML tags
-      return <span>{cleanContent(content).replace(/<[^>]*>/g, '')}</span>;
-    }
-  };
-  
-  // If we're in the browser, process the content with React elements
-  if (typeof document !== 'undefined') {
-    return processContent();
-  }
-  
-  // Fallback to dangerouslySetInnerHTML if running on server
-  return <div dangerouslySetInnerHTML={{ __html: cleanContent(content) }} />;
-}
-
-// Define the DocumentMessage component
-function DocumentMessage({ message }) {
-  // Check if this is a document generation status message
-  const isGenerationStatus = typeof message.content === 'string' && 
-    message.content.includes('document-generation-status');
-  
-  // If it's a generation status message, show a proper loading UI
-  if (isGenerationStatus) {
-    return (
-      <div className="flex items-center justify-center py-6">
-        <div className="flex flex-col items-center gap-3">
-          <div className="flex items-center gap-2">
-            <Loader2 className="h-5 w-5 animate-spin text-muted-foreground" />
-            <span className="text-sm text-muted-foreground font-medium">I'm generating your document now</span>
-          </div>
-          <p className="text-xs text-muted-foreground max-w-xs text-center">
-            This typically takes about 1 minute to complete.
-          </p>
-          <p className="text-xs text-muted-foreground max-w-xs text-center">
-            You'll receive a notification and the document will appear here when it's ready.
-          </p>
-        </div>
-      </div>
-    );
-  }
-  
-  // First check if message has metadata with document links
-  const documentLinks = message.metadata?.documentLinks;
-  const hasMetadataLinks = documentLinks && documentLinks.googleDocLink;
-  
-  // Then check for n8n links in the message content as a fallback
-  const extractedLinks = extractN8nLinks(message.content);
-  const hasExtractedLinks = extractedLinks.googleDocLink;
-  
-  // Check if the content already contains HTML links
-  const contentHasHtmlLinks = typeof message.content === 'string' && message.content.includes('<a href');
-  
-  // Get the Google Doc link, prioritizing metadata links
-  const docLink = documentLinks?.googleDocLink || extractedLinks.googleDocLink;
-  
-  return (
-    <div className="space-y-3">
-      <div>
-        {/* Check if content contains HTML and render appropriately */}
-        {message.content.includes('<a href') ? (
-          <HTMLContent content={message.content} />
-        ) : (
-          <MarkdownMessage content={message.content} />
-        )}
-      </div>
-      {/* Only show the separate link section if content doesn't already have HTML links */}
-      {docLink && !contentHasHtmlLinks && (
-        <div className="bg-green-50 border border-green-200 rounded-lg p-4 space-y-2">
-          <div className="flex items-center gap-2 text-green-800 font-medium">
-            <CheckCircle2 className="h-5 w-5" />
-            Your document is ready!
-          </div>
-          <div className="flex items-center gap-2 text-sm">
-            <FileText className="h-4 w-4 text-green-600" />
-            <a 
-              href={docLink} 
-              target="_blank" 
-              rel="noopener noreferrer"
-              className="text-blue-600 hover:text-blue-800 underline break-all"
-            >
-              View Google Doc
-            </a>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Add a component for rendering HTML landing pages
-function LandingPageMessage({ content }) {
-  const [showPreview, setShowPreview] = useState(false);
-  const [copied, setCopied] = useState(false);
-  
-  // Extract HTML code from the message content
-  const extractHTMLCode = (text) => {
-    // Look for HTML code blocks or complete HTML documents
-    const htmlMatch = text.match(/```html\n([\s\S]*?)\n```/) || 
-                     text.match(/```\n(<!DOCTYPE html[\s\S]*?<\/html>)\n```/) ||
-                     text.match(/(<!DOCTYPE html[\s\S]*?<\/html>)/);
-    
-    return htmlMatch ? htmlMatch[1] : null;
-  };
-
-  const htmlCode = extractHTMLCode(content);
-  
-  const copyToClipboard = async () => {
-    if (htmlCode) {
-      try {
-        await navigator.clipboard.writeText(htmlCode);
-        setCopied(true);
-        setTimeout(() => setCopied(false), 2000);
-      } catch (err) {
-        console.error('Failed to copy:', err);
-      }
-    }
-  };
-
-  const downloadHTML = () => {
-    if (htmlCode) {
-      const blob = new Blob([htmlCode], { type: 'text/html' });
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = 'landing-page.html';
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-    }
-  };
-
-  if (!htmlCode) {
-    // If no HTML code found, render as regular markdown
-    return <MarkdownMessage content={content} />;
-  }
-
-  return (
-    <div className="space-y-4">
-      {/* Regular message content without the HTML code block */}
-      <div>
-        <MarkdownMessage content={content.replace(/```html\n[\s\S]*?\n```/g, '').replace(/```\n<!DOCTYPE html[\s\S]*?<\/html>\n```/g, '').replace(/<!DOCTYPE html[\s\S]*?<\/html>/g, '').trim()} />
-      </div>
-      
-      {/* HTML Code Section */}
-      <div className="border rounded-lg p-4 bg-gray-50 dark:bg-gray-800">
-        <div className="flex items-center justify-between mb-3">
-          <h4 className="font-semibold text-sm">Landing Page HTML</h4>
-          <div className="flex gap-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => setShowPreview(!showPreview)}
-              className="text-xs"
-            >
-              {showPreview ? 'Hide Preview' : 'Show Preview'}
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={copyToClipboard}
-              className="text-xs"
-            >
-              {copied ? 'Copied!' : 'Copy HTML'}
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={downloadHTML}
-              className="text-xs"
-            >
-              <Download className="h-3 w-3 mr-1" />
-              Download
-            </Button>
-          </div>
-        </div>
-        
-        {showPreview && (
-          <div className="mb-4">
-            <div className="border rounded bg-white" style={{ height: '400px' }}>
-              <iframe
-                srcDoc={htmlCode}
-                className="w-full h-full rounded"
-                title="Landing Page Preview"
-                sandbox="allow-same-origin"
-              />
-            </div>
-          </div>
-        )}
-        
-        <div className="bg-gray-100 dark:bg-gray-900 rounded p-3 text-xs font-mono overflow-x-auto max-h-40 overflow-y-auto">
-          <pre className="whitespace-pre-wrap">{htmlCode}</pre>
-        </div>
-        
-        <div className="mt-3 text-xs text-gray-600 dark:text-gray-400">
-          <p><strong>Instructions:</strong></p>
-          <ol className="list-decimal list-inside space-y-1 mt-1">
-            <li>Copy the HTML code above</li>
-            <li>In HighLevel, go to Sites → Pages → Create New Page</li>
-            <li>Choose "Custom Code" or "Blank Page"</li>
-            <li>Paste the HTML code into the custom code section</li>
-            <li>Save and publish your landing page</li>
-          </ol>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Add a function to check if a message contains landing page HTML
-function isLandingPageMessage(message) {
-  if (typeof message.content !== 'string') return false;
-  
-  // Check if the message contains HTML code blocks or complete HTML documents
-  const hasHTMLCode = message.content.includes('```html') || 
-                     message.content.includes('<!DOCTYPE html') ||
-                     (message.content.includes('<html') && message.content.includes('</html>'));
-  
-  return hasHTMLCode;
-}
 
 export default function ChatArea({ selectedTool, currentChat, setCurrentChat, chats, setChats }) {
   const [input, setInput] = useState("");
@@ -437,8 +43,6 @@ export default function ChatArea({ selectedTool, currentChat, setCurrentChat, ch
   const [questionsAnswered, setQuestionsAnswered] = useState(0); // Add questionsAnswered state
   const [isInitiating, setIsInitiating] = useState(false);
   const [initiationAttemptedForContext, setInitiationAttemptedForContext] = useState(false);
-  const [isWaitingForN8n, setIsWaitingForN8n] = useState(false);
-  const eventSourceRef = useRef(null);
   const textareaRef = useRef(null);
   const scrollAreaRef = useRef(null);
   const prevChatIdRef = useRef();
@@ -446,6 +50,8 @@ export default function ChatArea({ selectedTool, currentChat, setCurrentChat, ch
   const { user } = useAuth();
   const lastMessageRef = useRef(null);
   const { track } = usePostHog();
+  const { connectToN8nResultStream, isWaitingForN8n, setIsWaitingForN8n, eventSourceRef } =
+    useChatSSE({ user, currentChat, setCurrentChat, chats, setChats, scrollToBottom, textareaRef });
 
   // Add this useEffect to track the isWaitingForN8n state
   useEffect(() => {
@@ -1182,510 +788,6 @@ export default function ChatArea({ selectedTool, currentChat, setCurrentChat, ch
                          (currentChat?.metadata?.questionsAnswered >= 6);
 
   // Function to connect to SSE endpoint
-  const connectToN8nResultStream = (chatId, encodedAnswers) => {
-    if (eventSourceRef.current) {
-      console.log("[SSE Connect] Closing existing EventSource before creating new one.");
-      eventSourceRef.current.close(); // Close any previous connection
-    }
-
-    // Get the last 30 messages from the current chat
-    let chatHistory = [];
-    if (currentChat && currentChat.messages) {
-      chatHistory = currentChat.messages.slice(-30).map(msg => ({
-        role: msg.role,
-        content: typeof msg.content === 'string' ? msg.content : JSON.stringify(msg.content) // Ensure content is string
-      }));
-    }
-
-    // Create the request body as a JSON object instead of URL params
-    const postData = {
-      chatId: chatId,
-      userId: user?.id || null,
-      answersData: JSON.parse(decodeURIComponent(encodedAnswers)), // Parse since we already have encoded JSON
-      chatHistory: chatHistory
-    };
-
-    console.log(`[SSE Connect] Connecting to /api/n8n-result with POST request`);
-    console.log(`[SSE Connect] POST data:`, {
-      chatId,
-      userId: user?.id || null,
-      answersDataFields: Object.keys(postData.answersData),
-      chatHistoryLength: chatHistory.length
-    });
-
-    // First, save an initial document generation message that will persist across refreshes
-    try {
-      if (user?.id) {
-        // Save both a text message for DB persistence and thread-level metadata to track generation state
-        const initialMessagePayload = {
-          thread_id: chatId,
-          role: 'assistant',
-          content: `
-          <div class="document-generation-status">
-            <p>📝 <strong>I'm generating your document now.</strong> Please wait...</p>
-            <p>This typically takes about 1 minute to complete.</p>
-            <p>You can safely refresh the page - the document will appear here when it's ready.</p>
-          </div>
-          `,
-          timestamp: new Date().toISOString(),
-          metadata: {
-            isGenerating: true,
-            generationStarted: new Date().toISOString()
-          }
-        };
-        console.log('[SSE Connect] Saving initial document generation message to DB:', initialMessagePayload);
-        saveMessage(initialMessagePayload, user.id)
-          .then(() => {
-            console.log('[SSE Connect] Successfully saved initial document message to DB.');
-            // Also update the thread metadata to indicate document generation is in progress
-            return fetch('/api/update-thread-metadata', {
-              method: 'POST',
-              headers: { 'Content-Type': 'application/json' },
-              body: JSON.stringify({
-                threadId: chatId,
-                metadata: {
-                  ...currentChat?.metadata,
-                  isGeneratingDocument: true,
-                  generationStartTime: new Date().toISOString()
-                }
-              })
-            });
-          })
-          .then(response => {
-            if (!response.ok) {
-              console.warn('[SSE Connect] Failed to update thread metadata:', response.status);
-            } else {
-              console.log('[SSE Connect] Successfully updated thread metadata for document generation');
-            }
-          })
-          .catch(err => console.error('[SSE Connect] Error in document generation setup:', err));
-          
-        // Also add to UI state to reflect message immediately - ENSURE IT GOES AT THE END
-        const initialMessage = { role: 'assistant', content: initialMessagePayload.content };
-        setChats(prevChats => prevChats.map(c => {
-          if (c.id === chatId) {
-            // Make sure we append to the end of the messages array
-            return {...c, messages: [...c.messages, initialMessage]};
-          }
-          return c;
-        }));
-        if (currentChat?.id === chatId) {
-          setCurrentChat(prevChat => ({...prevChat, messages: [...prevChat.messages, initialMessage]}));
-        }
-      } else {
-        console.warn('[SSE Connect] Cannot save initial document message - no user ID');
-      }
-    } catch (initErr) {
-      console.error('[SSE Connect] Error saving initial document message:', initErr);
-    }
-
-    // We need to use a custom implementation for EventSource with POST
-    // First, make the initial request to establish the connection
-    fetch('/api/n8n-result', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(postData)
-    })
-    .then(response => {
-      if (!response.ok) {
-        throw new Error(`SSE connection failed with status: ${response.status}`);
-      }
-      
-      // Create a reader for the response body stream
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder();
-      
-      // Track the buffer and last event info
-      let buffer = '';
-      
-      // Function to process SSE events from the buffer
-      const processEvents = (chunk) => {
-        buffer += chunk;
-        
-        // Process each event (separated by double newlines)
-        const events = buffer.split('\n\n');
-        buffer = events.pop() || '';
-        
-        events.forEach(eventStr => {
-          if (!eventStr.trim()) return;
-          
-          // Parse the event type and data
-          const eventLines = eventStr.split('\n');
-          let eventType = '';
-          let eventData = '';
-          
-          eventLines.forEach(line => {
-            if (line.startsWith('event:')) {
-              eventType = line.substring(6).trim();
-            } else if (line.startsWith('data:')) {
-              eventData = line.substring(5).trim();
-            }
-          });
-          
-          if (eventType && eventData) {
-            // Handle the event based on its type
-            try {
-              const parsedData = JSON.parse(eventData);
-              
-              if (eventType === 'n8n_result') {
-                handleN8nResult(parsedData);
-              } else if (eventType === 'error') {
-                handleErrorEvent(parsedData);
-              } else {
-                console.warn(`[SSE Connect] Unknown event type: ${eventType}`);
-              }
-            } catch (e) {
-              console.error(`[SSE Connect] Error parsing event data: ${e.message}`);
-            }
-          }
-        });
-      };
-      
-      // Handle n8n_result events
-      const handleN8nResult = (eventData) => {
-        console.log("[SSE Connect] Received n8n_result event:", JSON.stringify(eventData, null, 2));
-        let contentToSaveToDB = null;   // For saving to DB
-        let n8nResultData = null;
-
-        try {
-          if (eventData.success && eventData.data) {
-            n8nResultData = eventData.data; // Store for later use
-            console.log("[SSE Connect] Parsed n8n result data:", JSON.stringify(n8nResultData, null, 2));
-            
-            // Look for Google Doc link with different possible property names
-            const googleDocLink = n8nResultData.googleDocLink || n8nResultData.docUrl || n8nResultData.googleDocURL || n8nResultData.documentUrl;
-            
-            // Log the link to debug
-            console.log("[SSE Connect] Google Doc link extracted:", googleDocLink);
-            
-            if (googleDocLink) {
-              // Update n8nResultData with normalized link
-              n8nResultData = {
-                ...n8nResultData,
-                googleDocLink: googleDocLink
-              };
-              
-              // Construct plain text version with HTML link embedded directly in the content
-              contentToSaveToDB = `✅ Document generated successfully!\n\n<a href="${googleDocLink}" target="_blank" rel="noopener noreferrer">View Google Doc</a>\n\nLink: ${googleDocLink}`;
-            } else {
-              throw new Error('No Google Doc link found in the response.');
-            }
-          } else {
-            throw new Error(eventData.message || 'Received unsuccessful result from server.');
-          }
-        } catch (parseError) {
-          console.error("[SSE Connect] Error parsing n8n_result data or constructing message:", parseError);
-          contentToSaveToDB = "Document generated, but there was an issue displaying the link."; 
-        }
-
-        // Save the text version to DB (if content exists)
-        if (contentToSaveToDB && user?.id) {
-          try {
-            // Create a dedicated links object for the Google Doc only
-            const documentLinks = {
-              googleDocLink: n8nResultData?.googleDocLink
-            };
-            
-            console.log("[SSE Connect] Document link to save:", documentLinks);
-            
-            const messagePayload = {
-              thread_id: chatId,
-              role: 'assistant',
-              content: contentToSaveToDB,
-              timestamp: new Date().toISOString(),
-              metadata: {
-                documentLinks: documentLinks,
-                isGenerating: false,
-                generationCompleted: new Date().toISOString()
-              }
-            };
-            console.log('[SSE Connect] Saving n8n result message to DB:', JSON.stringify(messagePayload, null, 2));
-            
-            // Update thread metadata to indicate generation is complete
-            fetch('/api/update-thread-metadata', {
-              method: 'POST',
-              headers: { 'Content-Type': 'application/json' },
-              body: JSON.stringify({
-                threadId: chatId,
-                metadata: {
-                  ...currentChat?.metadata,
-                  isGeneratingDocument: false,
-                  documentGenerated: true,
-                  documentLinks: documentLinks,
-                  generationCompleteTime: new Date().toISOString()
-                }
-              })
-            }).then(response => {
-              if (!response.ok) {
-                console.warn('[SSE Connect] Failed to update thread metadata after completion:', response.status);
-              } else {
-                console.log('[SSE Connect] Successfully updated thread metadata for document completion');
-              }
-            }).catch(err => {
-              console.error('[SSE Connect] Error updating thread metadata:', err);
-            });
-            
-            // Call saveMessage and wait for it to complete
-            saveMessage(messagePayload, user.id)
-              .then(() => {
-                console.log('[SSE Connect] Successfully saved n8n result message to DB for thread:', chatId);
-                
-                // Update the UI for other open instances of this chat
-                const documentMessage = { 
-                  role: 'assistant', 
-                  content: contentToSaveToDB,
-                  metadata: {
-                    documentLinks: documentLinks
-                  }
-                };
-                
-                // Remove any "generating document" messages and add the new result message at the end
-                setCurrentChat(prevChat => {
-                  if (!prevChat || prevChat.id !== chatId) return prevChat;
-                  
-                  // Filter out any "generating document" messages first
-                  const filteredMessages = prevChat.messages.filter(m => 
-                    !(typeof m.content === 'string' && (
-                      m.content.includes("generating your document") || 
-                      m.content.includes("document-generation-status"))
-                    )
-                  );
-                  
-                  // Always add the new document message at the end
-                  return {...prevChat, messages: [...filteredMessages, documentMessage]};
-                });
-                
-                setChats(prevChats => prevChats.map(c => {
-                  if (c.id === chatId) {
-                    // Filter out any "generating document" messages
-                    const filteredMessages = c.messages.filter(m => 
-                      !(typeof m.content === 'string' && (
-                        m.content.includes("generating your document") || 
-                        m.content.includes("document-generation-status"))
-                      )
-                    );
-                    
-                    // Always add the new document message at the end
-                    return {...c, messages: [...filteredMessages, documentMessage]};
-                  }
-                  return c;
-                }));
-                
-                // Clear the waiting state
-                setIsWaitingForN8n(false);
-                
-                // Trigger scroll to show the new document message
-                setTimeout(() => {
-                  scrollToBottom();
-                }, 100);
-                
-                // TODO: Add notification system here
-                // This is where we would trigger a notification to let the user know
-                // their document is ready, even if they're not currently viewing this chat
-                console.log('[SSE Connect] Document ready - notification system would trigger here');
-              })
-              .catch(dbError => {
-                console.error('[SSE Connect] Error saving n8n result message to DB:', dbError);
-              });
-          } catch (dbError) {
-            console.error('[SSE Connect] Error preparing to save n8n result message to DB:', dbError);
-          }
-        } else {
-          console.warn(`[SSE Connect] Did not save n8n result message to DB. No user ID or content. contentExists=${!!contentToSaveToDB}, userId=${!!user?.id}`);
-        }
-      };
-      
-      // Handle error events
-      const handleErrorEvent = (eventData) => {
-        console.error("[SSE Connect] Received error event:", eventData);
-        
-        // Always save error to DB so it persists through refreshes
-        if (user?.id) {
-          const errorMessagePayload = {
-            thread_id: chatId,
-            role: 'assistant',
-            content: eventData.message || "Connection error while generating document. Please try again later.",
-            timestamp: new Date().toISOString()
-          };
-          
-          saveMessage(errorMessagePayload, user.id)
-            .then(() => console.log('[SSE Connect] Successfully saved error message to DB.'))
-            .catch(err => console.error('[SSE Connect] Error saving error message:', err));
-        }
-        
-        // Add an error message to the chat if we're on the relevant chat
-        const sseErrorMessage = { 
-          role: 'assistant', 
-          content: eventData.message || "Connection error while generating document. Please try again later.", 
-          isJSX: false
-        }; 
-        
-        if (currentChat?.id === chatId) {
-          setCurrentChat(prevChat => {
-            if (!prevChat) return null;
-            
-            // Remove any "generating document" messages
-            const filteredMessages = prevChat.messages.filter(m => 
-              m.content !== "I'm generating your document now. Please wait..."
-            );
-            
-            return {...prevChat, messages: [...filteredMessages, sseErrorMessage]};
-          });
-          
-          setChats(prevChats => prevChats.map(c => {
-            if (c.id === chatId) {
-              // Remove any "generating document" messages
-              const filteredMessages = c.messages.filter(m => 
-                m.content !== "I'm generating your document now. Please wait..."
-              );
-              return {...c, messages: [...filteredMessages, sseErrorMessage]};
-            }
-            return c;
-          }));
-        }
-        
-        setIsWaitingForN8n(false);
-        
-        // Trigger scroll to show the error message
-        setTimeout(() => {
-          scrollToBottom();
-        }, 100);
-        
-        textareaRef.current?.focus();
-      };
-      
-      // Function to read the next chunk
-      const readNextChunk = () => {
-        reader.read().then(({ done, value }) => {
-          if (done) {
-            console.log("[SSE Connect] Stream closed by server.");
-            setIsWaitingForN8n(false);
-            return;
-          }
-          
-          const chunk = decoder.decode(value, { stream: true });
-          processEvents(chunk);
-          readNextChunk(); // Continue reading
-        }).catch(error => {
-          console.error("[SSE Connect] Error reading from stream:", error);
-          setIsWaitingForN8n(false);
-          
-          // Save stream error to DB
-          if (user?.id) {
-            const streamErrorPayload = {
-              thread_id: chatId,
-              role: 'assistant',
-              content: "Error streaming document data. Please try again.",
-              timestamp: new Date().toISOString()
-            };
-            
-            saveMessage(streamErrorPayload, user.id)
-              .then(() => console.log('[SSE Connect] Successfully saved stream error message to DB.'))
-              .catch(err => console.error('[SSE Connect] Error saving stream error message:', err));
-          }
-          
-          // Add an error message to the chat
-          const streamErrorMessage = { 
-            role: 'assistant', 
-            content: "Error streaming document data. Please try again.", 
-            isJSX: false
-          };
-          
-          if (currentChat?.id === chatId) {
-            setCurrentChat(prevChat => {
-              if (!prevChat) return null;
-              
-              // Remove any "generating document" messages
-              const filteredMessages = prevChat.messages.filter(m => 
-                m.content !== "I'm generating your document now. Please wait..."
-              );
-              
-              return {...prevChat, messages: [...filteredMessages, streamErrorMessage]};
-            });
-            
-            setChats(prevChats => prevChats.map(c => {
-              if (c.id === chatId) {
-                // Remove any "generating document" messages
-                const filteredMessages = c.messages.filter(m => 
-                  m.content !== "I'm generating your document now. Please wait..."
-                );
-                return {...c, messages: [...filteredMessages, streamErrorMessage]};
-              }
-              return c;
-            }));
-          }
-          
-          // Trigger scroll to show the stream error message
-          setTimeout(() => {
-            scrollToBottom();
-          }, 100);
-          
-          textareaRef.current?.focus();
-        });
-      };
-      
-      // Start reading from the stream
-      readNextChunk();
-    })
-    .catch(error => {
-      console.error("[SSE Connect] Fetch error:", error);
-      setIsWaitingForN8n(false);
-      
-      // Save connection error to DB
-      if (user?.id) {
-        const connectionErrorPayload = {
-          thread_id: chatId,
-          role: 'assistant',
-          content: `Connection error: ${error.message}. Please try again later.`,
-          timestamp: new Date().toISOString()
-        };
-        
-        saveMessage(connectionErrorPayload, user.id)
-          .then(() => console.log('[SSE Connect] Successfully saved connection error message to DB.'))
-          .catch(err => console.error('[SSE Connect] Error saving connection error message:', err));
-      }
-      
-      // Add an error message to the chat
-      const connectionErrorMessage = { 
-        role: 'assistant', 
-        content: `Connection error: ${error.message}. Please try again later.`, 
-        isJSX: false
-      };
-      
-      if (currentChat?.id === chatId) {
-        setCurrentChat(prevChat => {
-          if (!prevChat) return null;
-          
-          // Remove any "generating document" messages
-          const filteredMessages = prevChat.messages.filter(m => 
-            m.content !== "I'm generating your document now. Please wait..."
-          );
-          
-          return {...prevChat, messages: [...filteredMessages, connectionErrorMessage]};
-        });
-        
-        setChats(prevChats => prevChats.map(c => {
-          if (c.id === chatId) {
-            // Remove any "generating document" messages
-            const filteredMessages = c.messages.filter(m => 
-              m.content !== "I'm generating your document now. Please wait..."
-            );
-            return {...c, messages: [...filteredMessages, connectionErrorMessage]};
-          }
-          return c;
-        }));
-      }
-      
-      // Trigger scroll to show the connection error message
-      setTimeout(() => {
-        scrollToBottom();
-      }, 100);
-      
-      textareaRef.current?.focus();
-    });
-  };
 
   const pollForAssistantResponse = async (threadId, runId, chatId, chatWithThinking, updatedMessages) => {
     console.log("[CHAT_DEBUG] Starting to poll for assistant response");
@@ -1814,176 +916,26 @@ export default function ChatArea({ selectedTool, currentChat, setCurrentChat, ch
           </div>
         </div>
       </div>
+        <MessageList
+          currentChat={currentChat}
+          user={user}
+          isWaitingForN8n={isWaitingForN8n}
+          isResponseLoading={isResponseLoading}
+          lastMessageRef={lastMessageRef}
+          scrollAreaRef={scrollAreaRef}
+        />
 
-      {/* Messages container - updated with responsive spacing */}
-      <ScrollArea 
-        ref={scrollAreaRef} 
-        className="flex-1 overflow-y-auto px-2 sm:px-4"
-      >
-        <div className="flex flex-col space-y-3 sm:space-y-6 py-4 sm:py-6 mb-16 sm:mb-20">
-          {/* First message or empty state when no messages */}
-          {!currentChat?.messages?.length ? (
-            <div className="flex items-center justify-center h-[calc(100vh-10rem)]">
-              <div className="text-center space-y-3 sm:space-y-6 max-w-md px-4">
-                <h3 className="text-lg sm:text-xl font-semibold">
-                  {selectedTool ? TOOLS[selectedTool].name : "Start a New Conversation"}
-                </h3>
-                <p className="text-sm sm:text-base text-muted-foreground">
-                  {selectedTool 
-                    ? TOOLS[selectedTool].description
-                    : "Ask me anything related to your business."}
-                </p>
-              </div>
-            </div>
-          ) : (
-            currentChat.messages
-              .filter((message) => {
-                // Filter out document generation status messages if document is already complete
-                const isGenerationStatus = typeof message.content === 'string' && 
-                  message.content.includes('document-generation-status');
-                
-                if (isGenerationStatus) {
-                  // Check if there are any completed document messages in the chat
-                  const hasCompletedDocuments = currentChat.messages.some(msg => 
-                    isDocumentMessage(msg) && !msg.content.includes('document-generation-status')
-                  );
-                  
-                  // Also check metadata for completion
-                  const isDocumentComplete = currentChat.metadata?.documentGenerated === true || 
-                    currentChat.metadata?.isGeneratingDocument === false;
-                  
-                  // Filter out generation status if document is complete
-                  if (hasCompletedDocuments || isDocumentComplete) {
-                    console.log('[ChatArea] Filtering out generation status message - document is complete');
-                    return false;
-                  }
-                }
-                
-                return true;
-              })
-              .map((message, index, filteredArray) => {
-                // Check if this is the last message in the filtered array
-                const isLastMessage = index === filteredArray.length - 1;
-                
-                return (
-                  <div
-                    key={message.id || `message-${index}`}
-                    className={`flex flex-col ${message.role === "user" ? "items-end" : "items-start"} `}
-                    ref={isLastMessage ? lastMessageRef : null}
-                  >
-                    {/* Message bubble - updated with responsive padding */}
-                    <div
-                      className={`
-                        flex items-start gap-2 sm:gap-3 max-w-[90%] sm:max-w-[85%] 
-                        ${message.role === "user" ? "flex-row-reverse" : ""}
-                      `}
-                    >
-                      {message.role === "user" ? (
-                        <Avatar className="h-8 w-8 sm:h-9 sm:w-9 mt-0.5">
-                          <AvatarImage src="" alt="User" />
-                          <div className="flex items-center justify-center h-full w-full bg-gradient-to-r from-purple-500 to-blue-500 text-white font-medium">
-                            {user?.email?.charAt(0).toUpperCase() || "U"}
-                          </div>
-                        </Avatar>
-                      ) : (
-                        <Avatar className="h-8 w-8 sm:h-9 sm:w-9 mt-0.5">
-                          <AvatarImage src="" alt="Assistant" />
-                          <div className="flex items-center justify-center h-full w-full bg-gradient-to-r from-blue-500 to-indigo-600 text-white font-medium">
-                            J
-                          </div>
-                        </Avatar>
-                      )}
+        <ChatInputForm
+          input={input}
+          setInput={setInput}
+          handleSubmit={handleSubmit}
+          handleKeyDown={handleKeyDown}
+          textareaRef={textareaRef}
+          isLoading={isLoading}
+          isResponseLoading={isResponseLoading}
+          isWaitingForN8n={isWaitingForN8n}
+        />
 
-                      <div
-                        className={`
-                          relative p-3 sm:p-4 rounded-lg text-sm sm:text-base space-y-1.5
-                          ${message.role === "user" ? "bg-primary text-primary-foreground" : "bg-muted"}
-                        `}
-                      >
-                        {message.is_thinking ? (
-                          <LoadingMessage content={message.content} role={message.role} />
-                        ) : (
-                          // Conditional rendering for different message types
-                          <>
-                            {isDocumentMessage(message) ? (
-                              <DocumentMessage message={message} />
-                            ) : isLandingPageMessage(message) ? (
-                              <LandingPageMessage content={message.content} />
-                            ) : (
-                              // Check for HTML content
-                              message.content.includes('<a href') ? (
-                                <HTMLContent content={message.content} />
-                              ) : (
-                                <MarkdownMessage content={message.content} />
-                              )
-                            )}
-                          </>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                );
-              })
-          )}
-
-          {/* Show n8n document generation loader */}
-          {isWaitingForN8n && (
-            <div className="flex items-center justify-center py-6">
-              <div className="flex flex-col items-center gap-3">
-                <div className="flex items-center gap-2">
-                  <Loader2 className="h-5 w-5 animate-spin text-muted-foreground" />
-                  <span className="text-sm text-muted-foreground font-medium">Generating document...</span>
-                </div>
-                <p className="text-xs text-muted-foreground max-w-xs text-center">
-                  This may take up to 1-3 minutes. Your document is being created based on your answers.
-                </p>
-              </div>
-            </div>
-          )}
-
-          {/* Loading state for AI response */}
-          {isResponseLoading && !isWaitingForN8n && (
-            <div className="flex items-start gap-3">
-              <Avatar className="h-9 w-9 mt-0.5">
-                <AvatarImage src="" alt="Assistant" />
-                <div className="flex items-center justify-center h-full w-full bg-gradient-to-r from-blue-500 to-indigo-600 text-white font-medium">
-                  J
-                </div>
-              </Avatar>
-              <div className="bg-muted p-4 rounded-lg">
-                <LoadingMessage role="assistant" />
-              </div>
-            </div>
-          )}
-        </div>
-      </ScrollArea>
-
-      {/* Input area - made responsive for mobile devices */}
-      <div className="absolute bottom-0 left-0 right-0 md:left-[300px] bg-background border-t p-3 sm:p-4">
-        <form onSubmit={handleSubmit} className="flex flex-col space-y-2 mobile-input-wrapper">
-          <div className="relative">
-            <Textarea
-              ref={textareaRef}
-              value={input}
-              onChange={(e) => setInput(e.target.value)}
-              onKeyDown={handleKeyDown}
-              placeholder="Type your message..."
-              className="resize-none pr-12 py-3 max-h-32 min-h-[52px] text-base font-medium mobile-input"
-              rows={1}
-              disabled={isLoading || isResponseLoading || isWaitingForN8n}
-              style={{ fontSize: '16px' }} /* Prevent iOS zoom by ensuring min 16px font */
-            />
-            <Button
-              type="submit"
-              size="icon"
-              className="absolute right-2 bottom-2 h-8 w-8 rounded-full"
-              disabled={!input.trim() || isLoading || isResponseLoading || isWaitingForN8n}
-            >
-              <ArrowUp className="h-4 w-4" />
-            </Button>
-          </div>
-        </form>
-      </div>
     </div>
   );
 }
