+++ components/chat/MarkdownMessage.js
@@ -1,0 +1,18 @@
+"use client";
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+
+export default function MarkdownMessage({ content }) {
+  return (
+    <ReactMarkdown
+      remarkPlugins={[remarkGfm]}
+      components={{
+        a: ({ node, ...props }) => (
+          <a {...props} target="_blank" rel="noopener noreferrer" className="text-blue-600 hover:underline" />
+        )
+      }}
+    >
+      {content}
+    </ReactMarkdown>
+  );
+}
